feat(ios): add openInSystemBrowser
This adds a new OSInAppBrowserInputArgumentsSystemBrowserModel structure that transforms all data to the required types requested by the library.
Create an error enumerator to deal with all errors.
Remove the External Browser RouterAdapter class as this now belongs to the library.
Add the default parameters for all Options.
Create an OpenInSystemBrowserParameterModel type with the method's parameters. This uses the one used for ExternalBrowser to get the URL.
Update the Example App to test the new feature.

References: https://outsystemsrd.atlassian.net/browse/RMET-3423
References: https://outsystemsrd.atlassian.net/browse/RMET-3429